+++ pyvisgen/gridding/__init__.py
@@ -1,11 +1,12 @@
-from .gridder import (
+from .gridder import ducc0_gridding, grid_data, grid_vis_loop_data
+from .utils import (
+    calc_truth_fft,
     convert_amp_phase,
     convert_real_imag,
-    ducc0_gridding,
-    grid_data,
-    grid_vis_loop_data,
+    create_gridded_data_set,
+    open_data,
+    save_fft_pair,
 )
-from .utils import calc_truth_fft, create_gridded_data_set, open_data, save_fft_pair
 
 __all__ = [
     "calc_truth_fft",
