+++ pyvisgen/gridding/utils.py
@@ -6,7 +6,7 @@ import numpy as np
 from tqdm.autonotebook import tqdm
 
 from pyvisgen.fits.data import fits_data
-from pyvisgen.gridding import convert_amp_phase, convert_real_imag, grid_data
+from pyvisgen.gridding import grid_data
 from pyvisgen.utils.config import read_data_set_conf
 from pyvisgen.utils.data import load_bundles, open_bundles
 
@@ -142,8 +142,8 @@ def save_fft_pair(path, x, y, name_x="x", name_y="y"):
     x = x[:, :, : half_image + 1, :]
     y = y[:, :, : half_image + 1, :]
 
-    test_shapes(x)
-    test_shapes(y)
+    test_shapes(x, "x")
+    test_shapes(y, "y")
 
     with h5py.File(path, "w") as hf:
         hf.create_dataset(name_x, data=x)
@@ -151,12 +151,18 @@ def save_fft_pair(path, x, y, name_x="x", name_y="y"):
         hf.close()
 
 
-def test_shapes(array):
+def test_shapes(array, name):
     if array.shape[1] != 2:
-        raise ValueError("Expected array axis 1 to be 2!")
+        raise ValueError(
+            f"Expected array {name} axis 1 to be 2 but got "
+            f"{array.shape} with axis 1: {array.shape[1]}!"
+        )
 
     if len(array.shape) != 4:
-        raise ValueError("Expected array shape to be of len 4!")
+        raise ValueError(
+            f"Expected array {name} shape to be of len 4 but got "
+            f"{array.shape} with len {len(array.shape)}!"
+        )
 
 
 def calc_truth_fft(sky_dist):
@@ -165,3 +171,32 @@ def calc_truth_fft(sky_dist):
     )
 
     return truth_fft
+
+
+def convert_amp_phase(data, sky_sim=False):
+    if sky_sim:
+        amp = np.abs(data)
+        phase = np.angle(data)
+        data = np.concatenate((amp, phase), axis=1)
+    else:
+        test = data[:, 0] + 1j * data[:, 1]
+        amp = np.abs(test)
+        phase = np.angle(test)
+        data = np.stack((amp, phase), axis=1)
+
+    return data
+
+
+def convert_real_imag(data, sky_sim=False):
+    if sky_sim:
+        real = data.real
+        imag = data.imag
+
+        data = np.concatenate((real, imag), axis=1)
+    else:
+        real = data[:, 0]
+        imag = data[:, 1]
+
+        data = np.stack((real, imag), axis=1)
+
+    return data
